Ingest: ingest IDs are ints, not longs.

+++ ArchivistSDK/src/main/java/com/zorroa/archivist/sdk/service/IngestService.java
@@ -32,7 +32,7 @@ public interface IngestService {
 
     boolean updateIngest(Ingest ingest, IngestUpdateBuilder builder);
 
-    Ingest getIngest(long id);
+    Ingest getIngest(int id);
 
     List<Ingest> getAllIngests();
